refactored RerouteTimeline's constants and functions
