+++ frontend/src/features/question_details/utils/formatDate.ts
@@ -1,0 +1,14 @@
+export const formatDate = (dateString?: string | Date | null): string => {
+  if (!dateString) return "";
+
+  const date =
+    typeof dateString === "string" ? new Date(dateString) : dateString;
+
+  if (isNaN(date.getTime())) return "";
+
+  return date.toLocaleDateString("en-US", {
+    month: "short",
+    day: "numeric",
+    year: "numeric",
+  });
+};
